MTurkManager no direct access to MTurkAgent state
